Refactor work-pools lib logic into component files
- Removed src/lib/work-pools.ts entirely
- Moved WorkPoolStatus type to shared src/components/work-pools/types.ts
- Moved getWorkPoolStatusInfo label logic to WorkPoolStatusBadge component as getStatusLabel
- Moved getWorkPoolStatusInfo description logic to WorkPoolStatusIcon component as getStatusDescription
- Removed unused isWorkPoolReady function
- Used switch statements for better TypeScript type safety
- All components now contain only the logic they actually use
- All tests passing, no functional changes
