+++ ui-v2/src/components/work-pools/types.ts
@@ -1,0 +1,1 @@
+export type WorkPoolStatus = "ready" | "paused" | "not_ready";
