+++ ui-v2/src/components/work-pools/work-pool-status-icon/work-pool-status-icon.tsx
@@ -7,7 +7,9 @@ import {
 	TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { cn } from "@/lib/utils";
-import { getWorkPoolStatusInfo, type WorkPoolStatus } from "@/lib/work-pools";
+import type { WorkPoolStatus } from "../types";
+
+export type { WorkPoolStatus };
 
 interface WorkPoolStatusIconProps {
 	status: WorkPoolStatus;
@@ -27,12 +29,23 @@ const statusColorMap = {
 	not_ready: "text-red-600",
 } as const;
 
+const getStatusDescription = (status: WorkPoolStatus): string => {
+	switch (status) {
+		case "ready":
+			return "Work pool is ready and accepting work";
+		case "paused":
+			return "Work pool is paused";
+		case "not_ready":
+			return "Work pool is not ready";
+	}
+};
+
 export const WorkPoolStatusIcon = ({
 	status,
 	showTooltip = true,
 	className,
 }: WorkPoolStatusIconProps) => {
-	const statusInfo = getWorkPoolStatusInfo(status);
+	const description = getStatusDescription(status);
 	const Icon = statusIconMap[status];
 	const colorClass = statusColorMap[status];
 
@@ -47,7 +60,7 @@ export const WorkPoolStatusIcon = ({
 			<Tooltip>
 				<TooltipTrigger asChild>{icon}</TooltipTrigger>
 				<TooltipContent>
-					<p>{statusInfo.description}</p>
+					<p>{description}</p>
 				</TooltipContent>
 			</Tooltip>
 		</TooltipProvider>
